Combine all the automatic data clearer tests into one test class
